# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MemContext API
  version: 1.1.0
  description: |
    REST API for persistent, evolving AI memory. Save, search, and manage
    memories with automatic deduplication, version chains, and hybrid retrieval.
servers:
  - url: https://api.memcontext.in
    description: Production
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/api-keys:
    get:
      tags: [API Keys]
      summary: List API keys
      operationId: listApiKeys
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Workspace-bound API keys owned by the signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListApiKeysResponse"
    post:
      tags: [API Keys]
      summary: Create API key
      operationId: createApiKey
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key created. The raw key is only returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApiKeyResponse"
  /api/api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Delete API key
      operationId: deleteApiKey
      security:
        - CookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: API key deleted.
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
  /api/subscription/current:
    get:
      tags: [Subscription]
      summary: Get workspace subscription
      operationId: getCurrentSubscription
      security:
        - CookieAuth: []
      parameters:
        - in: query
          name: workspaceId
          schema:
            type: string
            format: uuid
          description: Workspace to read. Defaults to the user's default workspace.
      responses:
        "200":
          description: Workspace subscription and member-memory usage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /api/subscription/checkout:
    post:
      tags: [Subscription]
      summary: Create workspace checkout session
      operationId: createCheckoutSession
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionPlanRequest"
      responses:
        "200":
          description: Checkout session URL. Only workspace owners can create one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
  /api/subscription/change-plan:
    post:
      tags: [Subscription]
      summary: Change workspace subscription plan
      operationId: changeSubscriptionPlan
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionPlanRequest"
      responses:
        "200":
          description: Plan change initiated. Only workspace owners can change billing.
          content:
            application/json:
              schema:
                type: object
                required: [success, message]
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/subscription/portal:
    post:
      tags: [Subscription]
      summary: Create billing portal session
      operationId: createBillingPortalSession
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingPortalRequest"
      responses:
        "200":
          description: Billing portal URL. Only workspace owners can manage billing.
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
                    format: uri
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key created from the MemContext dashboard. Keys are prefixed with `mc_`, bound to one workspace, and API-key requests cannot override that workspace.
    CookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Dashboard session cookie used for workspace/company endpoints.
...
  schemas:
...
    PlanType:
      type: string
      enum: [free, hobby, pro, ultimate]
    SubscriptionStatus:
      type: string
      enum: [active, cancelled, expired, on_hold, failed]
    Subscription:
      type: object
      required: [workspaceId, plan, memoryCount, memoryLimit, status]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        workspaceId:
          type: string
          format: uuid
        plan:
          $ref: "#/components/schemas/PlanType"
        memoryCount:
          type: integer
          minimum: 0
        memoryLimit:
          type: integer
          minimum: 0
        status:
          $ref: "#/components/schemas/SubscriptionStatus"
        renewsAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
    ApiKey:
      type: object
      required: [id, userId, workspaceId, keyPrefix, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        workspaceId:
          type: string
          format: uuid
        workspaceName:
          type: string
        keyPrefix:
          type: string
        name:
          type: string
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    ListApiKeysResponse:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
    CreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        workspaceId:
          type: string
          format: uuid
          description: Workspace to bind the key to. Defaults to the user's default workspace.
    CreateApiKeyResponse:
      type: object
      required: [id, name, keyPrefix, workspaceId, key, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
        workspaceId:
          type: string
          format: uuid
        key:
          type: string
          description: Raw API key. It is only returned once.
        createdAt:
          type: string
          format: date-time
    SubscriptionPlanRequest:
      type: object
      required: [plan, workspaceId]
      properties:
        plan:
          type: string
          enum: [hobby, pro, ultimate]
        workspaceId:
          type: string
          format: uuid
    CheckoutSessionResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        sessionId:
          type: string
    BillingPortalRequest:
      type: object
      required: [workspaceId]
      properties:
        workspaceId:
          type: string
          format: uuid
...